finished fitApp. logic of adding todays dishes, background of all VC's, pickers for height and weight

+++ fitApp/fitApp/Model/UserModel.swift
@@ -11,15 +11,15 @@ import Foundation
 class User{
     var name: String?
     var phoneNumber: String?
-    var currentWeight: Double?
-    var currentHeight: Double?
+    var currentWeight: Int?
+    var currentHeight: Int?
     var calories: Int?
     var isGetWeight: Bool?
     var period: Int?
     var plan: PlanModel?
     
     
-    init(name: String?, phoneNumber: String?, currentWeight: Double?, currentHeight: Double?, calories: Int?, isGetWeight: Bool?, period: Int?) {
+    init(name: String?, phoneNumber: String?, currentWeight: Int?, currentHeight: Int?, calories: Int?, isGetWeight: Bool?, period: Int?) {
         self.name = name
         self.phoneNumber = phoneNumber
         self.currentWeight = currentWeight
